+++ Player.cpp
@@ -1,7 +1,7 @@
 #include "Player.h"
 
 
-Player::Player(float newX, float newY, int newWidth, int newHeight) 
+Player::Player(int newX, int newY, int newWidth, int newHeight)
 	: Entity(newX, newY, newWidth, newHeight){}
 
 int Player::getLife() {
